Window System Refactor
Before:
Upon login in, a play button would enable to allow the player to press it and to load into the minigame.

Now:

Upon login in, the game will automatically load you into the minigame.
The game no longer allow the player to go back to the login menu as this is no longer required.